+++ feed/index.html
@@ -52,7 +52,6 @@
             <ul class="dropdown-menu" aria-labelledby="sortDropdown">
                 <li><a class="dropdown-item" href="#" id="sortNewest">Newest</a></li>
                 <li><a class="dropdown-item" href="#" id="sortOldest">Oldest</a></li>
-                <li><a class="dropdown-item" href="#" id="sortPopular">Most Popular</a></li>
             </ul>
         </div>
         <form class="mb-5" id="postForm">
@@ -85,7 +84,9 @@
         <p>&copy; 2023 Socially. All rights reserved.</p>
     </footer>
     <script src="/js/bootstrap.bundle.min.js"></script>
-    <script type="module" src="/js/feed.js" defer></script>
-    <script type="module" src="/js/postDetail.js" defer></script>
+    <script type="module" src="/js/feed.js"></script>
+    <script type="module" src="/js/postDetail.js"></script>
+    <script type="module" src="/js/uiService.js"></script>
+    <script type="module" src="/js/main.js"></script>
 </body>
 </html>
