+++ index.html
@@ -74,6 +74,9 @@
     </footer>
 <script type="module" src="/js/auth.js"></script>
 <script src="/js/bootstrap.bundle.min.js"></script>
+<script type="module" src="/js/uiService.js" defer></script>
+<script type="module" src="/js/postDetail.js" defer></script>
+<script type="module" src="/js/main.js" defer></script>
 </body>
 </html>
