+++ js/postService.js
@@ -1,5 +1,6 @@
 import { getAuthHeader } from './auth.js';
 import { API_BASE_URL } from '../js/constants.js';
+import { state } from './main.js';
 
 /**
  * Fetches posts based on the current filter, search query, and pagination offset.
@@ -73,9 +74,6 @@ export async function fetchPosts(limit, offset, searchQuery, currentFilter, tagF
     }
 }
 
-
-
-
 /**
  * Sends a request to update a post with new data.
  * @async
@@ -99,12 +97,12 @@ export async function updatePost(id, updatedPost) {
             throw new Error('Failed to update post');
         }
         console.log('Post updated successfully');
-        fetchPosts();
+        // Call fetchPosts with limit and currentOffset from state
+        fetchPosts(state.limit, state.currentOffset);
     } catch (error) {
         console.error('Error updating post:', error);
     }
 }
-
 /**
  * Sends a request to delete a post by ID.
  * @async
@@ -126,7 +124,9 @@ export async function deletePost(id) {
             throw new Error('Failed to delete post');
         }
         console.log('Post deleted successfully');
-        fetchPosts(); // Refresh the posts
+
+        // Call fetchPosts with limit and currentOffset from state
+        fetchPosts(state.limit, state.currentOffset);
 
     } catch (error) {
         console.error('Error deleting post:', error);
@@ -186,8 +186,118 @@ export async function createPost(event) {
         }
         const data = await response.json();
         console.log('Post created successfully:', data);
-        fetchPosts();
+        fetchPosts(state.limit, state.currentOffset);
     } catch (error) {
         console.error('Error creating post:', error);
+    }
+}
+
+/**
+ * Fetches all unique tags from all posts and populates the tag filter dropdown.
+ * @async
+ * @function
+ * @throws Will throw an error if the request to fetch posts fails.
+ */
+export async function fetchAllTags() {
+    const url = `${API_BASE_URL}/social/posts`;
+    const options = {
+        headers: getAuthHeader(),
+    };
+    const response = await fetch(url, options);
+    if (!response.ok) {
+        throw new Error('Failed to fetch posts');
+    }
+    const posts = await response.json();
+    // Extract all tags from all posts, flatten the array, and remove duplicates
+    const allTags = posts.flatMap(post => post.tags);
+    const uniqueTags = [...new Set(allTags)];
+
+    // Get the tagFilter dropdown and populate it with the unique tags
+    const tagFilterDropdown = document.getElementById('tagFilter');
+    uniqueTags.forEach(tag => {
+        const option = document.createElement('option');
+        option.value = tag;
+        option.textContent = tag;
+        tagFilterDropdown.appendChild(option);
+    });
+}
+/**
+ * Displays an array of posts on the page.
+ * @function
+ * @param {Array} posts - An array of post objects to display.
+ */
+export function displayPosts(posts) {
+    const postsContainer = document.getElementById('postsContainer');
+    if (postsContainer) {
+        postsContainer.innerHTML = '';
+        posts.forEach(post => {
+            const postCard = document.createElement('div');
+            postCard.classList.add('card', 'mb-4');
+            const postImage = document.createElement('img');
+            postImage.src = post.media || '/img/panda.jpg';
+            postImage.onerror = () => {
+                postImage.src = '/img/panda.jpg';
+            };
+            postImage.classList.add('card-img-top');
+            postImage.alt = '';
+            postCard.appendChild(postImage);
+            const postBody = document.createElement('div');
+            postBody.classList.add('card-body');
+            const postTitle = document.createElement('h5');
+            postTitle.classList.add('card-title');
+            postTitle.textContent = post.title;
+            postBody.appendChild(postTitle);
+            const postContent = document.createElement('p');
+            postContent.classList.add('card-text');
+            postContent.textContent = post.body;
+            postBody.appendChild(postContent);
+            const readMoreButton = document.createElement('a');
+            readMoreButton.href = `/postDetail/index.html?id=${post.id}`;
+            readMoreButton.classList.add('btn', 'btn-primary');
+            readMoreButton.textContent = 'Read More';
+            postBody.appendChild(readMoreButton);
+            const editButton = document.createElement('button');
+            editButton.textContent = 'Edit';
+            editButton.classList.add('btn', 'btn-warning', 'me-2');
+            editButton.onclick = () => editPost(post.id);
+            postBody.appendChild(editButton);
+            const deleteButton = document.createElement('button');
+            deleteButton.textContent = 'Delete';
+            deleteButton.classList.add('btn', 'btn-danger');
+            deleteButton.onclick = () => deletePost(post.id);
+            postBody.appendChild(deleteButton);
+            postCard.appendChild(postBody);
+            postsContainer.appendChild(postCard);
+        });
+    } else {
+        console.error('Container element not found');
+    }
+}
+/**
+ * Handles the editing of a post by prompting the user for a new title and content.
+ * If the user cancels the operation at any prompt, an alert is shown, and the operation is cancelled.
+ * @function
+ * @param {string} id - The ID of the post to edit.
+ */
+
+export function editPost(id) {
+    const title = prompt('Enter the new title:');
+    if (title === null) {
+        alert('Edit operation cancelled.');
+        return;
+    }
+    
+    const content = prompt('Enter the new content:');
+    if (content === null) {
+        alert('Edit operation cancelled.');
+        return;
+    }
+
+    if (title && content) {
+        const updatedPost = {
+            title: title,
+            body: content,
+        };
+        updatePost(id, updatedPost);
     }
 }
