+++ src/scss/index.css
@@ -1,10 +1,3 @@
-body {
-  background-color: #F5F5F5;
-  font-family: "Arial", sans-serif;
-  color: #333;
-  line-height: 1.6;
-}
-
 .container {
   max-width: 800px;
   background-color: white;
@@ -39,6 +32,11 @@ body {
 .navbar form button.btn, .navbar form button.custom-btn {
   height: 36px;
   line-height: 20px;
+}
+
+#nextPage, #prevPage {
+  display: block !important;
+  visibility: visible !important;
 }
 
 .custom-btn {
